Rename mixin filename
Since the file is already in the mixins/ dir the extra "mixin" in the filename seems redundant.
