+++ city_scrapers/spiders/wicks_city_lb.py
@@ -1,4 +1,4 @@
-from city_scrapers.mixins.wichita_city_mixin import WichitaCityMixin
+from city_scrapers.mixins.wichita_city import WichitaCityMixin
 
 
 class WicksCityLBSpider(WichitaCityMixin):
